+++ Code/Editor/EditorEngineProcessFramework/EngineProcess/Implementation/EngineProcessApp.cpp
@@ -1,7 +1,7 @@
 #include <EditorEngineProcessFrameworkPCH.h>
 
-#include <Core/ActorSystem/Actor2.h>
-#include <Core/ActorSystem/ActorManager2.h>
+#include <Core/ActorSystem/Actor.h>
+#include <Core/ActorSystem/ActorManager.h>
 #include <EditorEngineProcessFramework/EngineProcess/EngineProcessApp.h>
 #include <EditorEngineProcessFramework/EngineProcess/RemoteViewContext.h>
 #include <GameEngine/GameApplication/GameApplication.h>
