+++ Code/Engine/GameEngine/GameApplication/Implementation/GameApplicationBase.cpp
@@ -1,6 +1,6 @@
 #include <GameEnginePCH.h>
 
-#include <Core/ActorSystem/ActorManager2.h>
+#include <Core/ActorSystem/ActorManager.h>
 #include <Core/Input/InputManager.h>
 #include <Core/ResourceManager/ResourceManager.h>
 #include <Foundation/Communication/GlobalEvent.h>
